backend/requirements.txt
fastapi
uvicorn
websockets
sqlalchemy
web3
python-dotenv
psycopg2-binary